Change modal by location to gallery modal
